refactor(labrinth): move block/unblock routes to public api
