+++ apps/labrinth/src/routes/internal/mod.rs
@@ -67,8 +67,6 @@ pub fn config(cfg: &mut web::ServiceConfig) {
 	),
 	paths(
 		admin::count_download,
-		blocked_users::block_user,
-		blocked_users::unblock_user,
 		blocked_users::block_status,
 		admin::force_reindex,
 		admin::force_reindex_project,
